requirements.txt
pykrx
pykiwoom
PyQt5
pandas
numpy
pyyaml
schedule
FinanceDataReader
